Move in-memory service implementation
This change moves all code related to the in-memory implementation of
the `RegistryService` interface in its own package under
`internal/service/inmemory`. This is in preparation of the additional
db-based implementation of the same interface.

This change also drops the `DeploymentProvider` interface which won't
be necessary anymore.
