+++ docs/thv-registry-api/docs.go
@@ -6,10 +6,10 @@ import "github.com/swaggo/swag/v2"
 
 const docTemplate = `{
     "schemes": {{ marshal .Schemes }},
-    "components": {"schemas":{"github_com_stacklok_toolhive-registry-server_internal_service.DeployedServer":{"properties":{"endpoint_url":{"type":"string"},"image":{"type":"string"},"name":{"type":"string"},"namespace":{"type":"string"},"ready":{"type":"boolean"},"status":{"type":"string"},"transport":{"type":"string"}},"type":"object"},"internal_api_v0.EnvVarDetail":{"properties":{"default":{"type":"string"},"description":{"type":"string"},"name":{"type":"string"},"required":{"type":"boolean"},"secret":{"type":"boolean"}},"type":"object"},"internal_api_v0.ErrorResponse":{"properties":{"error":{"type":"string"}},"type":"object"},"internal_api_v0.ListServersResponse":{"properties":{"servers":{"items":{"$ref":"#/components/schemas/internal_api_v0.ServerSummaryResponse"},"type":"array","uniqueItems":false},"total":{"type":"integer"}},"type":"object"},"internal_api_v0.RegistryInfoResponse":{"properties":{"last_updated":{"type":"string"},"source":{"type":"string"},"total_servers":{"type":"integer"},"version":{"type":"string"}},"type":"object"},"internal_api_v0.ServerDetailResponse":{"properties":{"args":{"items":{"type":"string"},"type":"array","uniqueItems":false},"description":{"type":"string"},"env_vars":{"items":{"$ref":"#/components/schemas/internal_api_v0.EnvVarDetail"},"type":"array","uniqueItems":false},"image":{"type":"string"},"metadata":{"additionalProperties":{},"type":"object"},"name":{"type":"string"},"permissions":{"additionalProperties":{},"type":"object"},"repository_url":{"type":"string"},"status":{"type":"string"},"tags":{"items":{"type":"string"},"type":"array","uniqueItems":false},"tier":{"type":"string"},"tools":{"items":{"type":"string"},"type":"array","uniqueItems":false},"transport":{"type":"string"},"volumes":{"additionalProperties":{},"type":"object"}},"type":"object"},"internal_api_v0.ServerSummaryResponse":{"properties":{"description":{"type":"string"},"name":{"type":"string"},"status":{"type":"string"},"tier":{"type":"string"},"tools_count":{"type":"integer"},"transport":{"type":"string"}},"type":"object"},"model.Argument":{"properties":{"choices":{"items":{"type":"string"},"type":"array","uniqueItems":false},"default":{"type":"string"},"description":{"type":"string"},"format":{"$ref":"#/components/schemas/model.Format"},"isRepeated":{"type":"boolean"},"isRequired":{"type":"boolean"},"isSecret":{"type":"boolean"},"name":{"example":"--port","type":"string"},"placeholder":{"type":"string"},"type":{"$ref":"#/components/schemas/model.ArgumentType"},"value":{"type":"string"},"valueHint":{"example":"file_path","type":"string"},"variables":{"additionalProperties":{"$ref":"#/components/schemas/model.Input"},"type":"object"}},"type":"object"},"model.ArgumentType":{"example":"positional","type":"string","x-enum-varnames":["ArgumentTypePositional","ArgumentTypeNamed"]},"model.Format":{"type":"string","x-enum-varnames":["FormatString","FormatNumber","FormatBoolean","FormatFilePath"]},"model.Icon":{"properties":{"mimeType":{"example":"image/png","type":"string"},"sizes":{"items":{"type":"string"},"type":"array","uniqueItems":false},"src":{"example":"https://example.com/icon.png","format":"uri","maxLength":255,"type":"string"},"theme":{"type":"string"}},"type":"object"},"model.Input":{"properties":{"choices":{"items":{"type":"string"},"type":"array","uniqueItems":false},"default":{"type":"string"},"description":{"type":"string"},"format":{"$ref":"#/components/schemas/model.Format"},"isRequired":{"type":"boolean"},"isSecret":{"type":"boolean"},"placeholder":{"type":"string"},"value":{"type":"string"}},"type":"object"},"model.KeyValueInput":{"properties":{"choices":{"items":{"type":"string"},"type":"array","uniqueItems":false},"default":{"type":"string"},"description":{"type":"string"},"format":{"$ref":"#/components/schemas/model.Format"},"isRequired":{"type":"boolean"},"isSecret":{"type":"boolean"},"name":{"example":"SOME_VARIABLE","type":"string"},"placeholder":{"type":"string"},"value":{"type":"string"},"variables":{"additionalProperties":{"$ref":"#/components/schemas/model.Input"},"type":"object"}},"type":"object"},"model.Package":{"properties":{"environmentVariables":{"description":"EnvironmentVariables are set when running the package","items":{"$ref":"#/components/schemas/model.KeyValueInput"},"type":"array","uniqueItems":false},"fileSha256":{"description":"FileSHA256 is the SHA-256 hash for integrity verification (required for mcpb, optional for others)","example":"fe333e598595000ae021bd27117db32ec69af6987f507ba7a63c90638ff633ce","pattern":"^[a-f0-9]{64}$","type":"string"},"identifier":{"description":"Identifier is the package identifier:\n  - For NPM/PyPI/NuGet: package name or ID\n  - For OCI: full image reference (e.g., \"ghcr.io/owner/repo:v1.0.0\")\n  - For MCPB: direct download URL","example":"@modelcontextprotocol/server-brave-search","minLength":1,"type":"string"},"packageArguments":{"description":"PackageArguments are passed to the package's binary","items":{"$ref":"#/components/schemas/model.Argument"},"type":"array","uniqueItems":false},"registryBaseUrl":{"description":"RegistryBaseURL is the base URL of the package registry (used by npm, pypi, nuget; not used by oci, mcpb)","example":"https://registry.npmjs.org","format":"uri","type":"string"},"registryType":{"description":"RegistryType indicates how to download packages (e.g., \"npm\", \"pypi\", \"oci\", \"nuget\", \"mcpb\")","example":"npm","minLength":1,"type":"string"},"runtimeArguments":{"description":"RuntimeArguments are passed to the package's runtime command (e.g., docker, npx)","items":{"$ref":"#/components/schemas/model.Argument"},"type":"array","uniqueItems":false},"runtimeHint":{"description":"RunTimeHint suggests the appropriate runtime for the package","example":"npx","type":"string"},"transport":{"$ref":"#/components/schemas/model.Transport"},"version":{"description":"Version is the package version (required for npm, pypi, nuget; optional for mcpb; not used by oci where version is in the identifier)","example":"1.0.2","minLength":1,"type":"string"}},"type":"object"},"model.Repository":{"properties":{"id":{"example":"b94b5f7e-c7c6-d760-2c78-a5e9b8a5b8c9","type":"string"},"source":{"example":"github","type":"string"},"subfolder":{"example":"src/everything","type":"string"},"url":{"example":"https://github.com/modelcontextprotocol/servers","format":"uri","type":"string"}},"type":"object"},"model.Status":{"type":"string","x-enum-varnames":["StatusActive","StatusDeprecated","StatusDeleted"]},"model.Transport":{"description":"Transport is required and specifies the transport protocol configuration","properties":{"headers":{"items":{"$ref":"#/components/schemas/model.KeyValueInput"},"type":"array","uniqueItems":false},"type":{"example":"stdio","type":"string"},"url":{"example":"https://api.example.com/mcp","type":"string"}},"type":"object"},"v0.Metadata":{"properties":{"count":{"type":"integer"},"nextCursor":{"type":"string"}},"type":"object"},"v0.RegistryExtensions":{"properties":{"isLatest":{"type":"boolean"},"publishedAt":{"format":"date-time","type":"string"},"status":{"$ref":"#/components/schemas/model.Status"},"updatedAt":{"format":"date-time","type":"string"}},"type":"object"},"v0.ResponseMeta":{"properties":{"io.modelcontextprotocol.registry/official":{"$ref":"#/components/schemas/v0.RegistryExtensions"}},"type":"object"},"v0.ServerJSON":{"properties":{"$schema":{"example":"https://static.modelcontextprotocol.io/schemas/2025-10-17/server.schema.json","format":"uri","minLength":1,"type":"string"},"_meta":{"$ref":"#/components/schemas/v0.ServerMeta"},"description":{"example":"MCP server providing weather data and forecasts via OpenWeatherMap API","maxLength":100,"minLength":1,"type":"string"},"icons":{"items":{"$ref":"#/components/schemas/model.Icon"},"type":"array","uniqueItems":false},"name":{"example":"io.github.user/weather","maxLength":200,"minLength":3,"pattern":"^[a-zA-Z0-9.-]+/[a-zA-Z0-9._-]+$","type":"string"},"packages":{"items":{"$ref":"#/components/schemas/model.Package"},"type":"array","uniqueItems":false},"remotes":{"items":{"$ref":"#/components/schemas/model.Transport"},"type":"array","uniqueItems":false},"repository":{"$ref":"#/components/schemas/model.Repository"},"title":{"example":"Weather API","maxLength":100,"minLength":1,"type":"string"},"version":{"example":"1.0.2","type":"string"},"websiteUrl":{"example":"https://modelcontextprotocol.io/examples","format":"uri","type":"string"}},"type":"object"},"v0.ServerListResponse":{"properties":{"metadata":{"$ref":"#/components/schemas/v0.Metadata"},"servers":{"items":{"$ref":"#/components/schemas/v0.ServerResponse"},"type":"array","uniqueItems":false}},"type":"object"},"v0.ServerMeta":{"properties":{"io.modelcontextprotocol.registry/publisher-provided":{"additionalProperties":{},"type":"object"}},"type":"object"},"v0.ServerResponse":{"properties":{"_meta":{"$ref":"#/components/schemas/v0.ResponseMeta"},"server":{"$ref":"#/components/schemas/v0.ServerJSON"}},"type":"object"}}},
+    "components": {"schemas":{"internal_api_v0.EnvVarDetail":{"properties":{"default":{"type":"string"},"description":{"type":"string"},"name":{"type":"string"},"required":{"type":"boolean"},"secret":{"type":"boolean"}},"type":"object"},"internal_api_v0.ErrorResponse":{"properties":{"error":{"type":"string"}},"type":"object"},"internal_api_v0.ListServersResponse":{"properties":{"servers":{"items":{"$ref":"#/components/schemas/internal_api_v0.ServerSummaryResponse"},"type":"array","uniqueItems":false},"total":{"type":"integer"}},"type":"object"},"internal_api_v0.RegistryInfoResponse":{"properties":{"last_updated":{"type":"string"},"source":{"type":"string"},"total_servers":{"type":"integer"},"version":{"type":"string"}},"type":"object"},"internal_api_v0.ServerDetailResponse":{"properties":{"args":{"items":{"type":"string"},"type":"array","uniqueItems":false},"description":{"type":"string"},"env_vars":{"items":{"$ref":"#/components/schemas/internal_api_v0.EnvVarDetail"},"type":"array","uniqueItems":false},"image":{"type":"string"},"metadata":{"additionalProperties":{},"type":"object"},"name":{"type":"string"},"permissions":{"additionalProperties":{},"type":"object"},"repository_url":{"type":"string"},"status":{"type":"string"},"tags":{"items":{"type":"string"},"type":"array","uniqueItems":false},"tier":{"type":"string"},"tools":{"items":{"type":"string"},"type":"array","uniqueItems":false},"transport":{"type":"string"},"volumes":{"additionalProperties":{},"type":"object"}},"type":"object"},"internal_api_v0.ServerSummaryResponse":{"properties":{"description":{"type":"string"},"name":{"type":"string"},"status":{"type":"string"},"tier":{"type":"string"},"tools_count":{"type":"integer"},"transport":{"type":"string"}},"type":"object"},"model.Argument":{"properties":{"choices":{"items":{"type":"string"},"type":"array","uniqueItems":false},"default":{"type":"string"},"description":{"type":"string"},"format":{"$ref":"#/components/schemas/model.Format"},"isRepeated":{"type":"boolean"},"isRequired":{"type":"boolean"},"isSecret":{"type":"boolean"},"name":{"example":"--port","type":"string"},"placeholder":{"type":"string"},"type":{"$ref":"#/components/schemas/model.ArgumentType"},"value":{"type":"string"},"valueHint":{"example":"file_path","type":"string"},"variables":{"additionalProperties":{"$ref":"#/components/schemas/model.Input"},"type":"object"}},"type":"object"},"model.ArgumentType":{"example":"positional","type":"string","x-enum-varnames":["ArgumentTypePositional","ArgumentTypeNamed"]},"model.Format":{"type":"string","x-enum-varnames":["FormatString","FormatNumber","FormatBoolean","FormatFilePath"]},"model.Icon":{"properties":{"mimeType":{"example":"image/png","type":"string"},"sizes":{"items":{"type":"string"},"type":"array","uniqueItems":false},"src":{"example":"https://example.com/icon.png","format":"uri","maxLength":255,"type":"string"},"theme":{"type":"string"}},"type":"object"},"model.Input":{"properties":{"choices":{"items":{"type":"string"},"type":"array","uniqueItems":false},"default":{"type":"string"},"description":{"type":"string"},"format":{"$ref":"#/components/schemas/model.Format"},"isRequired":{"type":"boolean"},"isSecret":{"type":"boolean"},"placeholder":{"type":"string"},"value":{"type":"string"}},"type":"object"},"model.KeyValueInput":{"properties":{"choices":{"items":{"type":"string"},"type":"array","uniqueItems":false},"default":{"type":"string"},"description":{"type":"string"},"format":{"$ref":"#/components/schemas/model.Format"},"isRequired":{"type":"boolean"},"isSecret":{"type":"boolean"},"name":{"example":"SOME_VARIABLE","type":"string"},"placeholder":{"type":"string"},"value":{"type":"string"},"variables":{"additionalProperties":{"$ref":"#/components/schemas/model.Input"},"type":"object"}},"type":"object"},"model.Package":{"properties":{"environmentVariables":{"description":"EnvironmentVariables are set when running the package","items":{"$ref":"#/components/schemas/model.KeyValueInput"},"type":"array","uniqueItems":false},"fileSha256":{"description":"FileSHA256 is the SHA-256 hash for integrity verification (required for mcpb, optional for others)","example":"fe333e598595000ae021bd27117db32ec69af6987f507ba7a63c90638ff633ce","pattern":"^[a-f0-9]{64}$","type":"string"},"identifier":{"description":"Identifier is the package identifier:\n  - For NPM/PyPI/NuGet: package name or ID\n  - For OCI: full image reference (e.g., \"ghcr.io/owner/repo:v1.0.0\")\n  - For MCPB: direct download URL","example":"@modelcontextprotocol/server-brave-search","minLength":1,"type":"string"},"packageArguments":{"description":"PackageArguments are passed to the package's binary","items":{"$ref":"#/components/schemas/model.Argument"},"type":"array","uniqueItems":false},"registryBaseUrl":{"description":"RegistryBaseURL is the base URL of the package registry (used by npm, pypi, nuget; not used by oci, mcpb)","example":"https://registry.npmjs.org","format":"uri","type":"string"},"registryType":{"description":"RegistryType indicates how to download packages (e.g., \"npm\", \"pypi\", \"oci\", \"nuget\", \"mcpb\")","example":"npm","minLength":1,"type":"string"},"runtimeArguments":{"description":"RuntimeArguments are passed to the package's runtime command (e.g., docker, npx)","items":{"$ref":"#/components/schemas/model.Argument"},"type":"array","uniqueItems":false},"runtimeHint":{"description":"RunTimeHint suggests the appropriate runtime for the package","example":"npx","type":"string"},"transport":{"$ref":"#/components/schemas/model.Transport"},"version":{"description":"Version is the package version (required for npm, pypi, nuget; optional for mcpb; not used by oci where version is in the identifier)","example":"1.0.2","minLength":1,"type":"string"}},"type":"object"},"model.Repository":{"properties":{"id":{"example":"b94b5f7e-c7c6-d760-2c78-a5e9b8a5b8c9","type":"string"},"source":{"example":"github","type":"string"},"subfolder":{"example":"src/everything","type":"string"},"url":{"example":"https://github.com/modelcontextprotocol/servers","format":"uri","type":"string"}},"type":"object"},"model.Status":{"type":"string","x-enum-varnames":["StatusActive","StatusDeprecated","StatusDeleted"]},"model.Transport":{"description":"Transport is required and specifies the transport protocol configuration","properties":{"headers":{"items":{"$ref":"#/components/schemas/model.KeyValueInput"},"type":"array","uniqueItems":false},"type":{"example":"stdio","type":"string"},"url":{"example":"https://api.example.com/mcp","type":"string"}},"type":"object"},"v0.Metadata":{"properties":{"count":{"type":"integer"},"nextCursor":{"type":"string"}},"type":"object"},"v0.RegistryExtensions":{"properties":{"isLatest":{"type":"boolean"},"publishedAt":{"format":"date-time","type":"string"},"status":{"$ref":"#/components/schemas/model.Status"},"updatedAt":{"format":"date-time","type":"string"}},"type":"object"},"v0.ResponseMeta":{"properties":{"io.modelcontextprotocol.registry/official":{"$ref":"#/components/schemas/v0.RegistryExtensions"}},"type":"object"},"v0.ServerJSON":{"properties":{"$schema":{"example":"https://static.modelcontextprotocol.io/schemas/2025-10-17/server.schema.json","format":"uri","minLength":1,"type":"string"},"_meta":{"$ref":"#/components/schemas/v0.ServerMeta"},"description":{"example":"MCP server providing weather data and forecasts via OpenWeatherMap API","maxLength":100,"minLength":1,"type":"string"},"icons":{"items":{"$ref":"#/components/schemas/model.Icon"},"type":"array","uniqueItems":false},"name":{"example":"io.github.user/weather","maxLength":200,"minLength":3,"pattern":"^[a-zA-Z0-9.-]+/[a-zA-Z0-9._-]+$","type":"string"},"packages":{"items":{"$ref":"#/components/schemas/model.Package"},"type":"array","uniqueItems":false},"remotes":{"items":{"$ref":"#/components/schemas/model.Transport"},"type":"array","uniqueItems":false},"repository":{"$ref":"#/components/schemas/model.Repository"},"title":{"example":"Weather API","maxLength":100,"minLength":1,"type":"string"},"version":{"example":"1.0.2","type":"string"},"websiteUrl":{"example":"https://modelcontextprotocol.io/examples","format":"uri","type":"string"}},"type":"object"},"v0.ServerListResponse":{"properties":{"metadata":{"$ref":"#/components/schemas/v0.Metadata"},"servers":{"items":{"$ref":"#/components/schemas/v0.ServerResponse"},"type":"array","uniqueItems":false}},"type":"object"},"v0.ServerMeta":{"properties":{"io.modelcontextprotocol.registry/publisher-provided":{"additionalProperties":{},"type":"object"}},"type":"object"},"v0.ServerResponse":{"properties":{"_meta":{"$ref":"#/components/schemas/v0.ResponseMeta"},"server":{"$ref":"#/components/schemas/v0.ServerJSON"}},"type":"object"}}},
     "info": {"contact":{"url":"https://github.com/stacklok/toolhive"},"description":"{{escape .Description}}","license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"},"title":"{{.Title}}","version":"{{.Version}}"},
     "externalDocs": {"description":"","url":""},
-    "paths": {"/api/v0/registry/info":{"get":{"deprecated":true,"description":"Get registry metadata including version, last updated time, and total servers","parameters":[{"description":"Response format","in":"query","name":"format","schema":{"default":"toolhive","enum":["toolhive","upstream"],"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.RegistryInfoResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Bad Request"},"501":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Implemented"}},"summary":"Get registry information","tags":["registry"]}},"/api/v0/registry/openapi.yaml":{"get":{"deprecated":true,"description":"Returns the OpenAPI specification for the registry API in YAML format","responses":{"200":{"content":{"application/json":{"schema":{"type":"string"}},"application/x-yaml":{"schema":{"type":"string"}}},"description":"OpenAPI specification in YAML format"}},"summary":"Get OpenAPI specification","tags":["system"]}},"/api/v0/registry/servers":{"get":{"deprecated":true,"description":"Get a list of all available MCP servers in the registry","parameters":[{"description":"Response format","in":"query","name":"format","schema":{"default":"toolhive","enum":["toolhive","upstream"],"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ListServersResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Bad Request"},"501":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Implemented"}},"summary":"List all servers","tags":["servers"]}},"/api/v0/registry/servers/deployed":{"get":{"deprecated":true,"description":"Get a list of all currently deployed MCP servers","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/github_com_stacklok_toolhive-registry-server_internal_service.DeployedServer"},"type":"array"}}},"description":"OK"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"List deployed servers","tags":["deployed-servers"]}},"/api/v0/registry/servers/deployed/{name}":{"get":{"deprecated":true,"description":"Get all deployed MCP servers that match the specified server registry name","parameters":[{"description":"Server registry name","in":"path","name":"name","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/github_com_stacklok_toolhive-registry-server_internal_service.DeployedServer"},"type":"array"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Get deployed servers by registry name","tags":["deployed-servers"]}},"/api/v0/registry/servers/{name}":{"get":{"deprecated":true,"description":"Get detailed information about a specific MCP server","parameters":[{"description":"Server name","in":"path","name":"name","required":true,"schema":{"type":"string"}},{"description":"Response format","in":"query","name":"format","schema":{"default":"toolhive","enum":["toolhive","upstream"],"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ServerDetailResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Found"},"501":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Implemented"}},"summary":"Get server by name","tags":["servers"]}},"/extension/v0/registries":{"get":{"description":"List all registries","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"List registries","tags":["extension"]}},"/extension/v0/registries/{registryName}":{"delete":{"description":"Delete a registry","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Delete registry","tags":["extension"]},"get":{"description":"Get a registry by name","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Get registry","tags":["extension"]},"put":{"description":"Create or update a registry","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Create or update registry","tags":["extension"]}},"/extension/v0/registries/{registryName}/servers/{serverName}/versions/{version}":{"delete":{"description":"Delete a server from the registry","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded version to retrieve (e.g., \\","in":"path","name":"version","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Delete server","tags":["extension"]},"put":{"description":"Create or update a server in the registry","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded version to retrieve (e.g., \\","in":"path","name":"version","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Create or update server","tags":["extension"]}},"/health":{"get":{"deprecated":true,"description":"Check if the registry API is healthy","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"OK"}},"summary":"Health check","tags":["system"]}},"/readiness":{"get":{"deprecated":true,"description":"Check if the registry API is ready to serve requests","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"OK"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Service Unavailable"}},"summary":"Readiness check","tags":["system"]}},"/registry/v0.1/publish":{"post":{"description":"Publish a server to the registry","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Publish server","tags":["registry","official"]}},"/registry/v0.1/servers":{"get":{"description":"Get a list of available servers in the registry","parameters":[{"description":"Pagination cursor for retrieving next set of results","in":"query","name":"cursor","schema":{"type":"string"}},{"description":"Maximum number of items to return","in":"query","name":"limit","schema":{"type":"integer"}},{"description":"Search servers by name (substring match)","in":"query","name":"search","schema":{"type":"string"}},{"description":"Filter by version ('latest' for latest version, or an exact version like '1.2.3')","in":"query","name":"version","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerListResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"}},"summary":"List servers","tags":["registry","official"]}},"/registry/v0.1/servers/{serverName}/versions":{"get":{"description":"Returns all available versions for a specific MCP server, ordered by publication date (newest first)","parameters":[{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerListResponse"}}},"description":"A list of all versions for the server"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"404":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Server not found"}},"summary":"List all versions of an MCP server","tags":["registry","official"]}},"/registry/v0.1/servers/{serverName}/versions/{version}":{"get":{"description":"Returns detailed information about a specific version of an MCP server.\nUse the special version ` + "`" + `latest` + "`" + ` to get the latest version.","parameters":[{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded version to retrieve (e.g., \\","in":"path","name":"version","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerResponse"}}},"description":"Detailed server information"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"404":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Server or version not found"}},"summary":"Get specific MCP server version","tags":["registry","official"]}},"/registry/{registryName}/v0.1/publish":{"post":{"description":"Publish a server to the registry","parameters":[{"description":"Registry name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Publish server","tags":["registry","official"]}},"/registry/{registryName}/v0.1/servers":{"get":{"description":"Get a list of available servers in the registry","parameters":[{"description":"Registry name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"Pagination cursor for retrieving next set of results","in":"query","name":"cursor","schema":{"type":"string"}},{"description":"Maximum number of items to return","in":"query","name":"limit","schema":{"type":"integer"}},{"description":"Search servers by name (substring match)","in":"query","name":"search","schema":{"type":"string"}},{"description":"Filter by version ('latest' for latest version, or an exact version like '1.2.3')","in":"query","name":"version","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerListResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"}},"summary":"List servers","tags":["registry","official"]}},"/registry/{registryName}/v0.1/servers/{serverName}/versions":{"get":{"description":"Returns all available versions for a specific MCP server, ordered by publication date (newest first)","parameters":[{"description":"Registry name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerListResponse"}}},"description":"A list of all versions for the server"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"404":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Server not found"}},"summary":"List all versions of an MCP server","tags":["registry","official"]}},"/registry/{registryName}/v0.1/servers/{serverName}/versions/{version}":{"get":{"description":"Returns detailed information about a specific version of an MCP server.\nUse the special version ` + "`" + `latest` + "`" + ` to get the latest version.","parameters":[{"description":"Registry name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded version to retrieve (e.g., \\","in":"path","name":"version","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerResponse"}}},"description":"Detailed server information"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"404":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Server or version not found"}},"summary":"Get specific MCP server version","tags":["registry","official"]}},"/v0/publish":{"post":{"deprecated":true,"description":"Publish a new MCP server to the registry or update an existing one","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"501":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Implemented"}},"summary":"Publish MCP server (Not Implemented)","tags":["servers"]}},"/version":{"get":{"deprecated":true,"description":"Get version information about the registry API","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"OK"}},"summary":"Version information","tags":["system"]}}},
+    "paths": {"/api/v0/registry/info":{"get":{"deprecated":true,"description":"Get registry metadata including version, last updated time, and total servers","parameters":[{"description":"Response format","in":"query","name":"format","schema":{"default":"toolhive","enum":["toolhive","upstream"],"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.RegistryInfoResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Bad Request"},"501":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Implemented"}},"summary":"Get registry information","tags":["registry"]}},"/api/v0/registry/openapi.yaml":{"get":{"deprecated":true,"description":"Returns the OpenAPI specification for the registry API in YAML format","responses":{"200":{"content":{"application/json":{"schema":{"type":"string"}},"application/x-yaml":{"schema":{"type":"string"}}},"description":"OpenAPI specification in YAML format"}},"summary":"Get OpenAPI specification","tags":["system"]}},"/api/v0/registry/servers":{"get":{"deprecated":true,"description":"Get a list of all available MCP servers in the registry","parameters":[{"description":"Response format","in":"query","name":"format","schema":{"default":"toolhive","enum":["toolhive","upstream"],"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ListServersResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Bad Request"},"501":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Implemented"}},"summary":"List all servers","tags":["servers"]}},"/api/v0/registry/servers/{name}":{"get":{"deprecated":true,"description":"Get detailed information about a specific MCP server","parameters":[{"description":"Server name","in":"path","name":"name","required":true,"schema":{"type":"string"}},{"description":"Response format","in":"query","name":"format","schema":{"default":"toolhive","enum":["toolhive","upstream"],"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ServerDetailResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Found"},"501":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Implemented"}},"summary":"Get server by name","tags":["servers"]}},"/extension/v0/registries":{"get":{"description":"List all registries","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"List registries","tags":["extension"]}},"/extension/v0/registries/{registryName}":{"delete":{"description":"Delete a registry","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Delete registry","tags":["extension"]},"get":{"description":"Get a registry by name","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Get registry","tags":["extension"]},"put":{"description":"Create or update a registry","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Create or update registry","tags":["extension"]}},"/extension/v0/registries/{registryName}/servers/{serverName}/versions/{version}":{"delete":{"description":"Delete a server from the registry","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded version to retrieve (e.g., \\","in":"path","name":"version","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Delete server","tags":["extension"]},"put":{"description":"Create or update a server in the registry","parameters":[{"description":"Registry Name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded version to retrieve (e.g., \\","in":"path","name":"version","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Create or update server","tags":["extension"]}},"/health":{"get":{"deprecated":true,"description":"Check if the registry API is healthy","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"OK"}},"summary":"Health check","tags":["system"]}},"/readiness":{"get":{"deprecated":true,"description":"Check if the registry API is ready to serve requests","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"OK"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Service Unavailable"}},"summary":"Readiness check","tags":["system"]}},"/registry/v0.1/publish":{"post":{"description":"Publish a server to the registry","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Publish server","tags":["registry","official"]}},"/registry/v0.1/servers":{"get":{"description":"Get a list of available servers in the registry","parameters":[{"description":"Pagination cursor for retrieving next set of results","in":"query","name":"cursor","schema":{"type":"string"}},{"description":"Maximum number of items to return","in":"query","name":"limit","schema":{"type":"integer"}},{"description":"Search servers by name (substring match)","in":"query","name":"search","schema":{"type":"string"}},{"description":"Filter by version ('latest' for latest version, or an exact version like '1.2.3')","in":"query","name":"version","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerListResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"}},"summary":"List servers","tags":["registry","official"]}},"/registry/v0.1/servers/{serverName}/versions":{"get":{"description":"Returns all available versions for a specific MCP server, ordered by publication date (newest first)","parameters":[{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerListResponse"}}},"description":"A list of all versions for the server"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"404":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Server not found"}},"summary":"List all versions of an MCP server","tags":["registry","official"]}},"/registry/v0.1/servers/{serverName}/versions/{version}":{"get":{"description":"Returns detailed information about a specific version of an MCP server.\nUse the special version ` + "`" + `latest` + "`" + ` to get the latest version.","parameters":[{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded version to retrieve (e.g., \\","in":"path","name":"version","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerResponse"}}},"description":"Detailed server information"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"404":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Server or version not found"}},"summary":"Get specific MCP server version","tags":["registry","official"]}},"/registry/{registryName}/v0.1/publish":{"post":{"description":"Publish a server to the registry","parameters":[{"description":"Registry name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"501":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Not implemented"}},"summary":"Publish server","tags":["registry","official"]}},"/registry/{registryName}/v0.1/servers":{"get":{"description":"Get a list of available servers in the registry","parameters":[{"description":"Registry name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"Pagination cursor for retrieving next set of results","in":"query","name":"cursor","schema":{"type":"string"}},{"description":"Maximum number of items to return","in":"query","name":"limit","schema":{"type":"integer"}},{"description":"Search servers by name (substring match)","in":"query","name":"search","schema":{"type":"string"}},{"description":"Filter by version ('latest' for latest version, or an exact version like '1.2.3')","in":"query","name":"version","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerListResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"}},"summary":"List servers","tags":["registry","official"]}},"/registry/{registryName}/v0.1/servers/{serverName}/versions":{"get":{"description":"Returns all available versions for a specific MCP server, ordered by publication date (newest first)","parameters":[{"description":"Registry name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerListResponse"}}},"description":"A list of all versions for the server"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"404":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Server not found"}},"summary":"List all versions of an MCP server","tags":["registry","official"]}},"/registry/{registryName}/v0.1/servers/{serverName}/versions/{version}":{"get":{"description":"Returns detailed information about a specific version of an MCP server.\nUse the special version ` + "`" + `latest` + "`" + ` to get the latest version.","parameters":[{"description":"Registry name","in":"path","name":"registryName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded server name (e.g., \\","in":"path","name":"serverName","required":true,"schema":{"type":"string"}},{"description":"URL-encoded version to retrieve (e.g., \\","in":"path","name":"version","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/v0.ServerResponse"}}},"description":"Detailed server information"},"400":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Bad request"},"404":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"Server or version not found"}},"summary":"Get specific MCP server version","tags":["registry","official"]}},"/v0/publish":{"post":{"deprecated":true,"description":"Publish a new MCP server to the registry or update an existing one","requestBody":{"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"501":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/internal_api_v0.ErrorResponse"}}},"description":"Not Implemented"}},"summary":"Publish MCP server (Not Implemented)","tags":["servers"]}},"/version":{"get":{"deprecated":true,"description":"Get version information about the registry API","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"OK"}},"summary":"Version information","tags":["system"]}}},
     "openapi": "3.1.0",
     "tags": [
         {"description":"Registry metadata and information","name":"registry"},
